wsgi.py
import os
from dotenv import load_dotenv

dotenv_path = os.path.join(os.path.dirname(__file__), '.env')
if os.path.exists(dotenv_path):
    load_dotenv(dotenv_path)

from sennder.app import create_app, make_celery

app = create_app()
celery = make_celery(app)
